refactored AccountList into its own component for consistency
