+++ src/components/common/SessionAccounts.vue
@@ -1,36 +1,39 @@
 <template>
-  <div class="session-accounts">
-    <div class="session-accounts-top">
-      <h2>My accounts</h2>
-      <p>
-        You can use the accounts below to explore
-        <router-link to="/wallet">
-          Lunie
-        </router-link>
-        and to approve transactions.
-      </p>
-    </div>
-    <div v-for="account in accounts" :key="account.address">
-      <div class="card content">
-        <div class="content-left">
-          <h3>{{ account.name }}</h3>
-          <Bech32 :address="account.address" short-form />
+  <SessionFrame>
+    <div class="session">
+      <div class="session-container">
+        <h2 class="session-title">My accounts</h2>
+        <div class="session-main">
+          <p>
+            You can use the accounts below to explore
+            <router-link to="/wallet">
+              Lunie
+            </router-link>
+            and to approve transactions.
+          </p>
+
+          <AccountList :accounts="accounts" />
+
+          <div class="button-add-account">
+            <TmBtn value="Add Another Account" to="/accounts" color="primary" />
+          </div>
         </div>
       </div>
     </div>
-    <!-- <div class="button-add-account">
-      <TmBtn value="Add Account" color="primary" />
-    </div> -->
-  </div>
+  </SessionFrame>
 </template>
 
 <script>
-import TmBtn from "common/TmBtn"
+import AccountList from "common/AccountList"
 import Bech32 from "common/Bech32"
+import SessionFrame from "common/SessionFrame"
+import TmBtn from "common/TmBtn"
 export default {
   name: `session-accounts`,
   components: {
+    AccountList,
     Bech32,
+    SessionFrame,
     TmBtn
   },
   computed: {
@@ -46,47 +49,6 @@ export default {
 </script>
 
 <style scoped>
-.session-accounts {
-  padding: 2rem;
-  background: var(--fg);
-  border-left: 1px solid var(--bc-dim);
-}
-
-.session-accounts h2 {
-  color: var(--bright);
-  font-size: var(--h1);
-  font-weight: 500;
-  line-height: 3rem;
-}
-
-.card {
-  background: var(--app-nav);
-  border-radius: 2px;
-  padding: 1rem;
-  font-size: var(--m);
-  margin-bottom: 0.5rem;
-  border: 1px solid var(--bc);
-}
-
-.card h3 {
-  font-size: 14px;
-  font-weight: 400;
-}
-
-.content {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  padding-left: 1rem;
-}
-
-.content-left {
-  display: flex;
-  flex-direction: column;
-}
-
 .button-add-account {
   display: flex;
   justify-content: flex-end;
